requirements.txt
google-generativeai
python-dotenv
streamlit
streamlit-quill
